Make minor adjustments to user module

+++ src/main/java/nl/novi/backend_eindopdracht/repository/user/TrainerRepository.java
@@ -3,5 +3,5 @@ package nl.novi.backend_eindopdracht.repository.user;
 import nl.novi.backend_eindopdracht.model.user.Trainer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TrainerRepository extends JpaRepository<Trainer, Integer> {
+public interface TrainerRepository extends JpaRepository<Trainer, Long> {
 }
